refactor(span): add `oxc_str` crate and move `Atom`, `Ident`, `CompactStr` to it (#18387)
## Summary

- Add new `oxc_str` crate containing string types
- Move `Atom`, `Ident`, and `CompactStr` from `oxc_span` to `oxc_str`
- Re-export types from `oxc_span` for backwards compatibility

## Test plan

- [x] `cargo check` passes
- [x] `cargo test -p oxc_str -p oxc_span` passes
- [x] `just lint` passes
