+++ crates/oxc_str/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "oxc_span"
+name = "oxc_str"
 version = "0.110.0"
 authors.workspace = true
 categories.workspace = true
@@ -21,17 +21,14 @@ doctest = true
 
 [dependencies]
 oxc_allocator = { workspace = true }
-oxc_ast_macros = { workspace = true }
 oxc_estree = { workspace = true }
-oxc_str = { workspace = true }
 
 compact_str = { workspace = true }
-miette = { workspace = true }
 
 schemars = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 
 [features]
 default = []
-serialize = ["compact_str/serde", "dep:serde", "oxc_estree/serialize", "oxc_str/serialize"]
-schemars = ["dep:schemars", "oxc_str/schemars"]
+serialize = ["compact_str/serde", "dep:serde", "oxc_estree/serialize"]
+schemars = ["dep:schemars"]
